requirements.txt
flask>=3.0.0
pandas>=2.0.0
openpyxl>=3.1.0
fpdf2>=2.7.0
